+++ extensions/brew/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added new terminal options for Alacritty, kitty, WezTerm and Hyper.
 - Used app icons for actions to run commands in terminals.
 - Apply Destructive style to Uninstall actions.
+- Fallback to Terminal.App in case the selected terminal app is not installed.
 
 ## [Improvements] - 2024-06-14